+++ flexmeasures/data/models/reporting/aggregator.py
@@ -50,14 +50,16 @@ class AggregatorReporter(Reporter):
             belief_time = server_now()
 
         for input_description in input:
+            sensor = input_description["sensor"]
             # if name is not in belief_search_config, using the Sensor id instead
             column_name = input_description.get(
                 "name", f"sensor_{input_description['sensor'].id}"
             )
 
             source = input_description.get("source")
+            source = input_description.get("sources", source)
 
-            df = input_description["sensor"].search_beliefs(
+            df = sensor.search_beliefs(
                 event_starts_after=start,
                 event_ends_before=end,
                 resolution=resolution,
@@ -65,6 +67,12 @@ class AggregatorReporter(Reporter):
                 source=source,
                 one_deterministic_belief_per_event=True,
             )
+
+            # found multiple sources in the beliefs of df but no source is specified
+            if len(df.lineage.sources) > 1 and (source is None or len(source) == 0):
+                raise ValueError(
+                    "Missing attribute source or sources. The fields `source` or `sources` is required when having multiple sources within the time window."
+                )
 
             df = df.droplevel([1, 2, 3])